requirements_debug.txt
aenum==2.1.0
asn1crypto==0.24.0
astor==0.6.2
attrs==17.4.0
Automat==0.6.0
autopep8==1.3.5
base58==0.2.5
bitcoin==1.1.42
blessings==1.6.1
bpython==0.17.1
bumpversion==0.5.3
bytecode==0.5
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorlog==3.1.2
constantly==15.1.0
coverage==4.5.1
coveralls==1.3.0
coz-bytecode==0.5.1
cryptography==2.2.2
curtsies==0.3.0
cycler==0.10.0
docopt==0.6.2
ecdsa==0.13
Events==0.3
gevent==1.2.2
greenlet==0.4.13
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
klein==17.10.0
logzero==1.5.0
memory-profiler==0.52.0
mmh3==2.5.1
mock==2.0.0
mpmath==1.0.0
neo-boa==0.4.1
neo-python==0.6.8
neo-python-rpc==0.2.0
neocore==0.4.2
numpy==1.14.2
pbr==4.0.1
peewee==2.10.2
pluggy==0.6.0
plyvel==1.0.4
prompt-toolkit==1.0.15
psutil==5.4.3
py==1.5.3
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.6.1
Pygments==2.2.0
pymitter==0.2.3
Pympler==0.5
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
quantities==0.12.1
requests==2.18.4
scrypt==0.8.6
six==1.11.0
tqdm==4.20.0
Twisted==17.9.0
typing==3.6.4
urllib3==1.22
virtualenv==15.2.0
wcwidth==0.1.7
Werkzeug==0.14.1
zope.interface==4.4.3
